Refactoring.. better class naming
